refactor : AuthenticationInterceptor 추상화 리팩토링
