+++ src/main/java/nextstep/auth/model/authentication/interceptor/SessionAuthenticationInterceptor.java
@@ -1,5 +1,7 @@
-package nextstep.auth.model.authentication;
+package nextstep.auth.model.authentication.interceptor;
 
+import nextstep.auth.model.authentication.AuthenticationToken;
+import nextstep.auth.model.authentication.interceptor.AuthenticationInterceptor;
 import nextstep.auth.model.authentication.service.CustomUserDetailsService;
 import nextstep.auth.model.context.Authentication;
 import nextstep.auth.model.context.SecurityContext;
